requirements_complete.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Streamlit for frontend
streamlit==1.29.0

# Google AI packages
google-generativeai==0.8.5
google-api-core==2.25.1
google-auth==2.40.3

# Machine Learning and Text Processing
sentence-transformers==5.1.0
torch>=2.0.0
transformers>=4.40.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Document processing
pypdf==3.17.4
docx2txt==0.8

# Data validation and models
pydantic==2.5.0

# HTTP requests and utilities
requests==2.31.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Database (SQLite is built-in to Python)

# Additional utilities
typing-extensions>=4.8.0